💯test: fake 객체 생성 및 수정
- Travelogue 페이크 객체 생성자에서 member를 받지 않도록 수정
- SubTravelogue 페이크 객체 생성
