+++ src/test/java/shop/zip/travel/domain/post/fake/FakeSubTravelogue.java
@@ -1,0 +1,24 @@
+package shop.zip.travel.domain.post.fake;
+
+import shop.zip.travel.domain.post.subTravelogue.entity.SubTravelogue;
+
+public class FakeSubTravelogue extends SubTravelogue {
+
+  private final Long id;
+
+  public FakeSubTravelogue(Long id, SubTravelogue subTravelogue) {
+    super(
+        subTravelogue.getTitle(),
+        subTravelogue.getContent(),
+        subTravelogue.getAddresses(),
+        subTravelogue.getTransportationSet(),
+        subTravelogue.getPhotos()
+    );
+    this.id = id;
+  }
+
+  @Override
+  public Long getId() {
+    return id;
+  }
+}
