+++ src/test/java/shop/zip/travel/domain/post/fake/FakeTravelogue.java
@@ -1,0 +1,26 @@
+package shop.zip.travel.domain.post.fake;
+
+import shop.zip.travel.domain.post.travelogue.entity.Travelogue;
+
+public class FakeTravelogue extends Travelogue {
+
+	private final Long id;
+
+	public FakeTravelogue(Long id, Travelogue travelogue) {
+		super(
+				travelogue.getPeriod(),
+				travelogue.getTitle(),
+				travelogue.getCountry(),
+				travelogue.getThumbnail(),
+				travelogue.getCost(),
+				travelogue.getSubTravelogues(),
+				travelogue.getMember()
+		);
+		this.id = id;
+	}
+
+	@Override
+	public Long getId() {
+		return id;
+	}
+}
